[Refactor]: 접속한 사용자가 해당 채팅방에 존재하는지 확인하는 로직 추가 - 추가로, query string로처리하던 room id를 parameter로 변경
