+++ next/src/app/(home)/(communication)/channel/chat/[id]/delete.tsx
